finished deleting user, added logging out

+++ src/components/ProfileOptions.tsx
@@ -1,11 +1,13 @@
-import { useState, RefObject, useEffect } from "react";
+import { useState, RefObject} from "react";
 import useOutsideClick from "../customHooks/useOutsideClick";
 import { UserReduxI } from "../interfaces/userI";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../store/store";
 import { serverUrl } from "../utilities/URLs";
 import { uiActions } from "../store/uiSlice";
+import { userActions } from "../store/userSlice";
 import { modalActions } from "../store/modalSlice";
+import { useNavigate } from "react-router-dom";
 
 const ProfileOptions: React.FC<{
   userId: string;
@@ -13,6 +15,7 @@ const ProfileOptions: React.FC<{
   const dispatch = useDispatch();
   const user: UserReduxI = useSelector((state: RootState) => state.user);
   const ui = useSelector((state: RootState) => state.ui);
+  const navigate = useNavigate()
 
   const [showOptions, setShowOptions] = useState(false);
 
@@ -28,8 +31,12 @@ const ProfileOptions: React.FC<{
 
   const handleDelete = async () => {
     const token = localStorage.getItem("token");
+    if(user._id !== userId){
+      dispatch(uiActions.setError("only owner of the account can delete it"))
+      return
+    }
     try {
-      const response = await fetch(serverUrl + `/comments/${userId}`, {
+      const response = await fetch(serverUrl + `/user/delete/${userId}`, {
         method: "DELETE",
         headers: {
           "Content-Type": "application/json",
@@ -37,7 +44,11 @@ const ProfileOptions: React.FC<{
         },
       });
       if (!response.ok) {
-        console.error("deleting comment failed");
+        console.error("deleting user failed");
+      }
+      if(response.ok){
+        dispatch(userActions.logOut())
+        navigate('/')
       }
     } catch (err) {
       console.error(err);
@@ -46,7 +57,7 @@ const ProfileOptions: React.FC<{
 
   const handleDeleteClick = () => {
     handleDelete().catch(() => {
-      console.error("Failed to delete post");
+      console.error("Failed to delete user");
     });
   };
